refactor cram-console so that ball is in a separate file
