+++ services/cram-console/Cargo.toml
@@ -23,6 +23,8 @@ modals = { path = "../modals", features = ["cramium-soc"] }
 gam = { path = "../gam" }
 locales = { path = "../../locales" }
 usb-device-xous = { path = "../usb-device-xous" }
+cramium-hal = { path = "../../libs/cramium-hal", features = ["std"] }
+cram-hal-service = { path = "../cram-hal-service" }
 
 [features]
 "cramium-fpga" = ["utralib/cramium-fpga", "xous-pl230/cramium-fpga"]
